+++ src/App.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import Landing from './components/layout/Landing';
+import Routes from './components/routing/Routes';
 
 import './App.css';
 
@@ -12,6 +13,9 @@ const App = () => {
                 <Switch>
                     <Route exact path='/'>
                         <Landing />
+                    </Route>
+                    <Route>
+                        <Routes />
                     </Route>
                 </Switch>
             </>