+++ src/components/contact/ContactItem.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+const ContactItem = ({ type, icon, link }) => {
+    return (
+        <a
+            href={link}
+            target='_blank'
+            rel='noopener noreferrer external'
+            className='contact-item'
+        >
+            <p className='p'>{type}</p>
+            <FontAwesomeIcon icon={icon} />
+        </a>
+    );
+};
+
+export default ContactItem;